refactor: file structure cleanup for equipment
