+++ src/components/charsheet/equipment/armor/ArmorHelpContent.tsx
