+++ src/components/charsheet/equipment/cyber/cyberViewTypes.ts
