+++ src/components/charsheet/equipment/gear/GearDetail.tsx
