+++ src/components/charsheet/equipment/weapons/WeaponDetail.tsx
