+++ src/components/charsheet/equipment/weapons/WeaponHelpContent.tsx
